Face-Detection-Attendance-System-main/requirements.txt
# Core ML and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Face Recognition (without TensorFlow dependency)
facenet-pytorch>=2.5.0

# Data Processing and Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Image Processing
imageio>=2.31.0

# Utilities
tqdm>=4.65.0
python-dateutil>=2.8.0

# Testing and Development
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter for development (optional)
jupyter>=1.0.0
ipython>=8.0.0

# HTTP requests
requests>=2.31.0
urllib3>=2.0.0 